up libs

+++ util/build_ext_libs.py
@@ -71,7 +71,8 @@ def create_build_dir(src_dir: str):
     build_dir = os.path.normpath(os.path.join(ext_build_dir(), '__' + Path(src_dir).name))
     if src_dir.endswith('ITK'):
         build_dir = os.path.normpath(os.path.join(ext_build_dir(), '_I'))  # ITK windows build dir length limit
-    shutil.rmtree(build_dir, ignore_errors=False, onerror=handleRemoveReadonly)
+    if os.path.exists(build_dir):
+        shutil.rmtree(build_dir, ignore_errors=False, onerror=handleRemoveReadonly)
     os.mkdir(build_dir)
     return build_dir
 
@@ -1040,12 +1041,9 @@ def build_folly(src_dir: str, install_dir: str):
         orig_file4 = os.path.join(src_dir, 'CMakeLists.txt')
         bak_file4 = patch_file(orig_file4,
                                from_texts=[r'project(${PACKAGE_NAME} CXX C)',
-                                           r'${FOLLY_DIR}/experimental/JSONSchemaTester.cpp',
                                            ],
                                to_texts=['project(${PACKAGE_NAME} CXX C)\n'
                                          'set(CMAKE_FIND_LIBRARY_SUFFIXES .lib .a ${CMAKE_FIND_LIBRARY_SUFFIXES})\n',
-                                         '${FOLLY_DIR}/experimental/JSONSchemaTester.cpp\n'
-                                         '${FOLLY_DIR}/portability/Filesystem.cpp\n',
                                          ])
 
         if is_windows():
